[native] Extract removeScreensFromStack
