+++ native/utils/navigation-utils.js
@@ -102,9 +102,54 @@ function findRouteIndexWithKey(state: NavigationState, key: string): ?number {
   return null;
 }
 
+// This function walks from the back of the stack and calls filterFunc on each
+// screen until the stack is exhausted or filterFunc returns "break". A screen
+// will be removed if and only if filterFunc returns "remove" (not "break").
+function removeScreensFromStack<S: NavigationState>(
+  state: S,
+  filterFunc: (route: NavigationRoute) => 'keep' | 'remove' | 'break',
+): S {
+  const newRoutes = [];
+  let newIndex = state.index;
+  let screenRemoved = false;
+  let breakActivated = false;
+  for (let i = state.routes.length - 1; i >= 0; i--) {
+    const route = state.routes[i];
+    if (breakActivated) {
+      newRoutes.unshift(route);
+      continue;
+    }
+    const result = filterFunc(route);
+    if (result === 'break') {
+      breakActivated = true;
+    }
+    if (breakActivated || result === 'keep') {
+      newRoutes.unshift(route);
+      continue;
+    }
+    screenRemoved = true;
+    if (newIndex >= i) {
+      invariant(
+        newIndex !== 0,
+        'Attempting to remove current route and all before it',
+      );
+      newIndex--;
+    }
+  }
+  if (!screenRemoved) {
+    return state;
+  }
+  return {
+    ...state,
+    index: newIndex,
+    routes: newRoutes,
+  };
+}
+
 export {
   assertNavigationRouteNotLeafNode,
   getThreadIDFromParams,
   currentLeafRoute,
   findRouteIndexWithKey,
+  removeScreensFromStack,
 };
